Module iroh_metrics::core

Expose core types and traits

Structs

Core

Core is the base metrics struct. It manages the mapping between the metrics name and the actual metrics. It also carries a single prometheus registry to be used by all metrics.

Counter

Open Metrics Counter to measure discrete events.

UsageReporter

Exposes a simple API to report usage statistics.

UsageStatsReport

Usage statistics report.

Traits

HistogramType

Interface for all distribution based metrics.

Metric

Description of a group of metrics.

MetricType

Interface for all single value based metrics.

Type Aliases

UsageResource

Type alias for a metered resource.