Trait iroh_metrics::core::Metric

pub trait Metric: Default + Iterable + Sized + Debug + 'static + Send + Sync {
    // Required method
    fn name() -> &'static str;

    // Provided methods
    fn new(registry: &mut Registry) -> Self { ... }
    fn with_metric<T, F: FnOnce(&Self) -> T>(f: F) { ... }
    fn try_get() -> Option<&'static Self> { ... }
}

Description of a group of metrics.

Required Methods

fn name() -> &'static str

The name of this metric group.

Provided Methods

fn new(registry: &mut Registry) -> Self

Initializes this metric group.

fn with_metric<T, F: FnOnce(&Self) -> T>(f: F)

Access to this metrics group to record a metric. Only records if this metric is registered in the global registry.

fn try_get() -> Option<&'static Self>

Attempts to get the current metric from the global registry.

Object Safety

This trait is not object safe.