iroh_metrics

Module iterable

Traits for iterating over the fields of structs.

Structs

FieldIter

Iterator over the fields of a struct.

Traits

IntoIterable

Helper trait to convert from self to dyn Iterable.

Iterable

Trait for iterating over the fields of a struct.

Derive Macros

Iterable

Derives Iterable for a struct.