iroh_metrics::service
Trait MetricsSource

pub trait MetricsSource: Send + 'static {
    // Required method
    fn encode_openmetrics(&self) -> Result<String, Error>;
}

Available on crate feature service only.

Helper trait to abstract over different ways to access metrics.

Required Methods

fn encode_openmetrics(&self) -> Result<String, Error>

Encodes all metrics into a string in the Open Metrics text format.