iroh_metrics::service
Type Alias RwLockRegistry

pub type RwLockRegistry = Arc<RwLock<Registry>>;

Available on crate feature service only.

A cloneable Registry in a read-write lock.

Useful if you need mutable access to a registry, while also using the services defined in crate::service.

Aliased Type

struct RwLockRegistry { /* private fields */ }