iroh_metrics::static_core
Struct Core

pub struct Core { /* private fields */ }

Available on crate feature static_core only.

Core is the base metrics struct.

It manages the mapping between the metrics name and the actual metrics. It also carries a single prometheus registry to be used by all metrics.

Implementations

impl Core

pub fn init<F: FnOnce(&mut Registry, &mut ErasedSyncSet)>(f: F)

Must only be called once to init metrics.

Panics if called a second time.

pub fn try_init<F: FnOnce(&mut Registry, &mut ErasedSyncSet)>( f: F, ) -> Result<()>

Trieds to init the metrics.

pub fn get() -> Option<&'static Self>

Returns a reference to the core metrics.

pub fn registry(&self) -> &Registry

Available on crate feature metrics only.

Returns a reference to the prometheus registry.

pub fn get_collector<T: MetricsGroup>(&self) -> Option<&T>

Returns a reference to the mapped metrics instance.

pub fn encode_openmetrics(&self) -> String

Available on crate feature metrics only.

Encodes the current metrics registry to a string in the prometheus text exposition format.