iroh_metrics

Struct Registry

pub struct Registry { /* private fields */ }

A registry for MetricsGroup.

Implementations

impl Registry

pub fn sub_registry_with_prefix( &mut self, prefix: impl Into<Cow<'static, str>>, ) -> &mut Self

Creates a subregistry where all metrics are prefixed with prefix.

Returns a mutable reference to the subregistry.

pub fn sub_registry_with_labels( &mut self, labels: impl IntoIterator<Item = (impl Into<Cow<'static, str>>, impl Into<Cow<'static, str>>)>, ) -> &mut Self

Creates a subregistry where all metrics are labeled.

Returns a mutable reference to the subregistry.

pub fn sub_registry_with_label( &mut self, key: impl Into<Cow<'static, str>>, value: impl Into<Cow<'static, str>>, ) -> &mut Self

Creates a subregistry where all metrics have a key=value label.

Returns a mutable reference to the subregistry.

pub fn register(&mut self, metrics_group: Arc<dyn MetricsGroup>)

Registers a MetricsGroup into this registry.

pub fn register_all(&mut self, metrics_group_set: &impl MetricsGroupSet)

Registers a MetricsGroupSet into this registry.

pub fn register_all_prefixed( &mut self, metrics_group_set: &impl MetricsGroupSet, )

Registers a MetricsGroupSet into this registry, prefixing all metrics with the group set’s name.

pub fn encode_openmetrics_to_writer(&self, writer: &mut impl Write) -> Result

Encodes all metrics in the OpenMetrics text format.

This does not write the terminal # EOF\n string to writer. You can use encode_openmetrics_eof to do that.

pub fn schema_version(&self) -> u64

Returns the current schema version of this registry.

pub fn encode_schema(&self, schema: &mut Schema)

Encodes the schema of all registered metrics into the provided schema builder.

pub fn encode_values(&self, values: &mut Values)

Encodes the current values of all registered metrics into the provided values builder.