iroh_metrics
Trait Iterable

pub trait Iterable {
    // Required methods
    fn field_count(&self) -> usize;
    fn field(&self, n: usize) -> Option<(&'static str, &dyn Any)>;
}

Trait for iterating over the fields of a struct.

Required Methods

fn field_count(&self) -> usize

Returns the number of fields in the struct.

fn field(&self, n: usize) -> Option<(&'static str, &dyn Any)>

Returns the field name and dyn reference to the field.

Implementations

impl dyn Iterable

pub fn iter(&self) -> FieldIter<'_>

Returns an iterator over the fields of the struct.