iroh_metrics
Trait MetricsGroup

pub trait MetricsGroup:
    Any
    + Iterable
    + IntoIterable
    + Debug
    + 'static
    + Send
    + Sync {
    // Required method
    fn name(&self) -> &'static str;

    // Provided methods
    fn register(&self, registry: &mut Registry) { ... }
    fn describe(&self) -> DecribeIter<'_>  { ... }
    fn values(&self) -> ValuesIter<'_>  { ... }
}

Description of a group of metrics.

Required Methods

fn name(&self) -> &'static str

The name of this metric group.

Provided Methods

fn register(&self, registry: &mut Registry)

Available on crate feature metrics only.

Initializes this metric group.

fn describe(&self) -> DecribeIter<'_>

Returns an iterator over all metric items with their descriptions.

fn values(&self) -> ValuesIter<'_>

Returns an iterator over all metric items with their values and types.