iroh_metrics
Trait EncodeLabelSet

pub trait EncodeLabelSet:
    Hash
    + Eq
    + Clone
    + Send
    + Sync
    + 'static {
    // Required method
    fn encode_label_pairs(&self) -> Vec<LabelPair<'_>>;
}

Trait for types that can be encoded as a set of labels.

Implement this for label structs to use with Family. The struct must also implement Clone + Hash + Eq + Send + Sync.

Required Methods

fn encode_label_pairs(&self) -> Vec<LabelPair<'_>>

Returns the labels as key-value pairs.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.